Reorganize ranking agent/trigger functions on backend
