+++ functions/src/stages/ranking.triggers.ts
@@ -1,4 +1,3 @@
-import {Value} from '@sinclair/typebox/value';
 import {
   ParticipantProfileExtended,
   RankingStageParticipantAnswer,
@@ -12,6 +11,8 @@ import {
   LAS_WTL_STAGE_ID,
   ElectionStrategy,
 } from '@deliberation-lab/utils';
+import {addParticipantAnswerToRankingStagePublicData} from './ranking.utils';
+import {getFirestoreParticipant, getFirestoreStage} from '../utils/firestore';
 
 import * as admin from 'firebase-admin';
 import * as functions from 'firebase-functions';
@@ -30,77 +31,22 @@ export const updateRankingStagePublicData = onDocumentWritten(
   async (event) => {
     const data = event.data.after.data() as StageParticipantAnswer | undefined;
 
-    const stageDocument = app
-      .firestore()
-      .collection('experiments')
-      .doc(event.params.experimentId)
-      .collection('stages')
-      .doc(event.params.stageId);
-    const stage = (await stageDocument.get()).data() as StageConfig;
+    const stage = await getFirestoreStage(
+      event.params.experimentId,
+      event.params.stageId,
+    );
     if (stage.kind !== StageKind.RANKING) return;
 
-    const participantDocument = app
-      .firestore()
-      .collection('experiments')
-      .doc(event.params.experimentId)
-      .collection('participants')
-      .doc(event.params.participantId);
+    const participant = await getFirestoreParticipant(
+      event.params.experimentId,
+      event.params.participantId,
+    );
 
-    // Run document write as transaction to ensure consistency
-    await app.firestore().runTransaction(async (transaction) => {
-      // Get participant
-      const participant = (
-        await participantDocument.get()
-      ).data() as ParticipantProfileExtended;
-
-      const publicDocument = app
-        .firestore()
-        .collection('experiments')
-        .doc(event.params.experimentId)
-        .collection('cohorts')
-        .doc(participant.currentCohortId)
-        .collection('publicStageData')
-        .doc(event.params.stageId);
-
-      // For hardcoded WTL stage in LAS game only
-      const wtlDoc = app
-        .firestore()
-        .collection('experiments')
-        .doc(event.params.experimentId)
-        .collection('cohorts')
-        .doc(participant.currentCohortId)
-        .collection('publicStageData')
-        .doc(LAS_WTL_STAGE_ID);
-      // Update public stage data (current participant rankings, current winner)
-      const publicStageData = (
-        await publicDocument.get()
-      ).data() as RankingStagePublicData;
-      publicStageData.participantAnswerMap[participant.publicId] =
-        data.rankingList;
-
-      // Calculate rankings
-      let participantAnswerMap = publicStageData.participantAnswerMap;
-
-      // If experiment has hardcoded WTL stage (for LAS game), use the WTL
-      // stage/question IDs to only consider top ranking participants
-      const wtlResponse = await wtlDoc.get();
-      if (wtlResponse.exists) {
-        const wtlData = wtlResponse.data() as SurveyStagePublicData;
-
-        if (wtlData?.kind === StageKind.SURVEY) {
-          const candidateList = getRankingCandidatesFromWTL(wtlData);
-          participantAnswerMap = filterRankingsByCandidates(
-            participantAnswerMap,
-            candidateList,
-          );
-        }
-      }
-
-      // Calculate winner (not used in frontend if strategy is none)
-      publicStageData.winnerId =
-        getCondorcetElectionWinner(participantAnswerMap);
-
-      transaction.set(publicDocument, publicStageData);
-    });
+    addParticipantAnswerToRankingStagePublicData(
+      event.params.experimentId,
+      stage,
+      participant,
+      data,
+    );
   },
 );
